expired

+++ app/Filament/Resources/EventResource.php
@@ -431,7 +431,7 @@ class EventResource extends Resource
                                     ),
 
                                 Forms\Components\TextInput::make('price')
-                                    ->integer()
+                                    ->numeric()
                                     ->prefix(function (Forms\Get $get) {
                                         $currencyCode = $get('currency_code_id');
 
@@ -445,7 +445,7 @@ class EventResource extends Resource
                                 Forms\Components\TextInput::make('ticket_fee')
                                     ->label('Ticket fee (Online)')
                                     ->required()
-                                    ->integer()
+                                    ->numeric()
                                     ->helperText('This fee will be added to the ticket sale price which are bought online, put 0 to disable additional fees for tickets which are bought online, does not apply for free tickets, will be applied to future orders')
                                     ->prefix(function (Forms\Get $get) {
                                         $currencyCode = $get('currency_code_id');
@@ -460,7 +460,7 @@ class EventResource extends Resource
                                 Forms\Components\TextInput::make('promotionalprice')
                                     ->label('Promotional price')
                                     ->helperText('Set a price lesser than than the original price to indicate a promotion (this price will be the SALE price)')
-                                    ->integer()
+                                    ->numeric()
                                     ->prefix(function (Forms\Get $get) {
                                         $currencyCode = $get('currency_code_id');
